requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
ansible==8.3.0
ansible-core==2.15.2
async-timeout==5.0.1
attrs==25.1.0
Babel==2.8.0
bcrypt==3.2.0
blinker==1.4
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==3.4.1
cli-helpers==2.2.1
click==8.0.3
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
cryptography==3.4.8
dbus-python==1.2.18
discord.py==2.5.0
distro==1.7.0
distro-info==1.1+ubuntu0.1
frozenlist==1.5.0
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
jeepney==0.7.1
Jinja2==3.0.3
jmespath==0.10.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==2.0.1
more-itertools==8.10.0
multidict==6.1.0
netifaces==0.11.0
ntlm-auth==1.4.0
oauthlib==3.2.0
packaging==21.3
paramiko==2.9.3
pendulum==2.1.2
pgcli==3.3.1
pgspecial==1.11.10
praw==7.8.1
prawcore==2.4.0
prompt-toolkit==3.0.28
propcache==0.2.1
psycopg2==2.9.2
pygame==2.6.1
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pykerberos==1.1.14
PyNaCl==1.5.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.1
python-dotenv==1.0.1
pytz==2022.1
pytzdata==2020.1
pywinrm==0.3.0
PyYAML==5.4.1
requests==2.32.3
requests-kerberos==0.12.0
requests-ntlm==1.1.0
resolvelib==0.8.1
SecretStorage==3.3.1
setproctitle==1.2.2
six==1.16.0
sqlparse==0.4.2
systemd-python==234
tabulate==0.8.9
terminaltables==3.1.0
typing_extensions==4.12.2
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
update-checker==0.18.0
urllib3==1.26.5
wadllib==1.3.6
wcwidth==0.2.5
websocket-client==1.8.0
xmltodict==0.12.0
yarl==1.18.3
zipp==1.0.0
